Added AWS infrastructure setup with Terraform
This commit introduces a new Terraform configuration for setting up an AWS environment. The setup includes the creation of a VPC, two subnets within that VPC, and an EC2 instance. It also retrieves the latest Amazon Linux 2 AMI to be used in the EC2 instance. Additionally, it creates an IAM policy with full permissions. All resources are tagged appropriately for easy identification and management.