Updated Pulumi code and README
In the Pulumi code, removed the creation of a new IAM policy and instead looked up an existing one. The existing policy is then attached to the role. Also, renamed the EC2 instance for better identification.

In the README file, added a link to Resource Events in Dynatrace documentation under host problems section.