Added AWS flow logs and updated security settings
In this update, we've added AWS IAM role for flow logs and a CloudWatch log group to capture VPC flow logs. We also created an AWS Flow Log resource to monitor all traffic in the VPC.

We made changes to the subnet configurations - removed public IP mapping on launch for one of the subnets.

Security group settings were updated as well. The ingress rule now only allows SSH access from a specific IP instead of open access, and descriptions were added to both ingress and egress rules for better clarity.